Add BatchSlice and parse_batch_slice to batch_filter.py
